change the Couchbase settings
