+++ packages/msteams/package.json
@@ -14,7 +14,6 @@
     "test": "mocha --require @babel/register"
   },
   "dependencies": {
-    "@microsoft/microsoft-graph-client": "^1.3.0",
     "amqplib": "^0.5.3",
     "apollo-server-express": "^2.3.1",
     "botbuilder": "3.15.0",
